Ver.1.7.0

+++ Source/ShidenCore/Private/Expression/ShidenExpressionEvaluator.cpp
@@ -8,7 +8,7 @@ bool FShidenExpressionValue::TryToNumeric(float& OutValue, FString& ErrorMessage
 	switch (Type)
 	{
 	case EShidenExpressionValueType::Integer:
-		OutValue = static_cast<double>(IntValue);
+		OutValue = static_cast<float>(IntValue);
 		return true;
 
 	case EShidenExpressionValueType::Float:
@@ -16,7 +16,7 @@ bool FShidenExpressionValue::TryToNumeric(float& OutValue, FString& ErrorMessage
 		return true;
 
 	case EShidenExpressionValueType::Boolean:
-		OutValue = BoolValue ? 1.0 : 0.0;
+		OutValue = BoolValue ? 1.0f : 0.0f;
 		return true;
 
 	default:
